requirements.txt
Pillow==9.3.0
PySimpleGUI==4.60.4
